Tests/LibWeb/Text/input/HTML/Location-set-invalid-href-url.html

<script src="../include.js"></script>
<script>
    test(() => {
        try {
            location.href = 'http://@:www.invalid-url.com';
        } catch (e) {
            println(`Error setting href: ${e}`);
        }
    });
</script>
