Tests/LibWeb/Text/expected/HTML/Location-set-invalid-href-url.txt

Error setting href: SyntaxError: Invalid URL 'http://@:www.invalid-url.com'
